work on more stuff

+++ src/core/scene.go
@@ -4,6 +4,6 @@ import "github.com/hajimehoshi/ebiten/v2"
 type Scene interface {
 	Load()
 	Unload()
-	Update(float32)
+	Update(float64)
 	Draw(*ebiten.Image)
 }
